refactor: as is 패키지를 포함한 레거시 코드 삭제
