+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/handleradapter/HandlerAdapter.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe.handleradapter;
+package com.interface21.webmvc.servlet.mvc.handleradapter;
 
 import com.interface21.webmvc.servlet.ModelAndView;
